requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Middleware
starlette==0.27.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
loguru==0.7.2
